The Ray distributed genome assembled was added
This patch adds Ray v2.1.0 to CloudBioLinux. Ray is a
distributed genome and metagenome assembler. Ray is built on top
of RayPlatform and scales to very large numbers of cores. Ray is
licensed under the GNU General Public License version 3 and is
in C++ 1998. RayPlatform is licensed under the GNU Lesser General
Public License version 3, is in C++ 1998, and uses the Message
Passing Interface standard version 1.0 from the Message Passing
Interface Forum.

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>